+++ ISBI/VDIFStream.h
@@ -10,7 +10,7 @@
 #include <ctime>
 
 
-static constexpr int16_t DECODER_LEVEL_2BIT[] = { -3, -1, 1, 3 };
+static constexpr int8_t DECODER_LEVEL_2BIT[] = { -3, -1, 1, 3 };
 
 struct VDIFHeader {
   // Word 0
@@ -36,7 +36,7 @@ struct VDIFHeader {
   int32_t dataSize() const;
   int32_t samplesPerFrame() const;
   int32_t numberOfChannels() const;
-  void decode2bit(const std::vector<char>& frame, std::vector<int16_t>& out) const;
+  void decode2bit(const std::vector<char>& frame, std::vector<int8_t>& out) const;
 
   friend std::ostream& operator<<(std::ostream& os, const VDIFHeader& header) {
     os << "----- VDIF HEADER -----" << std::endl;
